backend/requirements.txt
Django>=5.0,<6.0
psycopg[binary]>=3.1,<4.0
Pillow>=10.0,<12.0
djangorestframework>=3.15